% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_CI.R
\name{check_CI}
\alias{check_CI}
\title{Check if zero is within the Credible intervals}
\usage{
check_CI(x, prob = 0.95)
}
\arguments{
\item{x}{vector holding the parameter draw from a MCMC of a MCMC or HMCMC sampler}

\item{probability}{quantile for the Creidble interval. Function only support 90 or 95 percent intervals.}
}
\value{
a charchter string that indicate whether the effect is signifcant or not
}
\description{
Check if zero is within the Credible intervals
}
\note{
From a Bayesian point of view this a very bad idea... use at your own risk.
}
\examples{
foo <- rnorm(1e4, 0.653, 0.192)
check_CI(foo, prob=0.95)
check_CI(bob, prob=0.90)
# apply on a matrix
bar <- matrix(rnorm(1e3*10,0.653, 0.192), ncol=10)
apply(bar, 2, check_CI, prob=0.95)
#END
}
