% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_interval.R
\name{check_interval}
\alias{check_interval}
\title{Look if a value is within an interval}
\usage{
check_interval(x, limit.A, limit.B)
}
\arguments{
\item{x}{value to test}

\item{limit.A}{lower limit}

\item{limit.B}{upper limit}
}
\description{
Look if a value is within an interval
}
