% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_km.R
\name{coords_to_km}
\alias{coords_to_km}
\title{Geographic coordinates to kilometers coordinates}
\usage{
coords_to_km(lon, lat, lon0, lat0)
}
\arguments{
\item{lon}{longitude vector in degrees}

\item{lat}{latitude vector in degrees}

\item{lon0}{longitude of the centroid of the survey area}

\item{lat0}{latittude of the centroid of the survey area}
}
\value{
Return a data frame holding the new coordinates
}
\description{
Transform geographic coordinates into a distance based coordinates. The function is based on the Haversine formula.
}
\examples{
deg.grid <- expand.grid(x=seq(40,50,by=2.5), y=seq(60,70,by=2.5))
km_.grid <- coords_to_km(lon=deg.grid$x, lat=deg.grid$y, lon0=mean(deg.grid$x), lat0=mean(deg.grid$y))
plot(km.grid)
#END
}

