% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_aniso_gaus.R
\name{corr_aniso}
\alias{corr_aniso}
\title{Gaussian anisotropic correlation function}
\usage{
corr_aniso(eta = 0.001, phi1 = 5, phi2 = 2.5, theta = 0,
  coords = expand.grid(x = seq(-10, 10, by = 2), y = seq(-10, 10, by = 2)))
}
\arguments{
\item{eta}{smoothness parameters. Control how smooth the correlation function is.}

\item{phi1}{decay parameter. Control the strength of the correlation between the points.}

\item{phi2}{decay parameter. Control the strength of the correlation between the points.}

\item{theta}{angle of the anisotropic field.}

\item{coords}{matrix holding the coordinates of the points. Remember that the points must be projected if you want a correaltion matrix that make sense.}
}
\value{
A Gaussian anisotropic correlation matrix between points.
}
\description{
Gaussian anisotropic correlation function
}
