% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_aniso_gaus.R
\name{covar_aniso}
\alias{covar_aniso}
\title{Gaussian anisotropic covariance function}
\usage{
covar_aniso(sigma = 1.25, eta = 0.001, phi1 = 5, phi2 = 2.5,
  theta = 0, coords = expand.grid(x = seq(-10, 10, by = 2), y = seq(-10, 10,
  by = 2)))
}
\arguments{
\item{sigma}{standard deviation for the covariance function}

\item{eta}{smoothness parameters. Control how smooth the covariance function is.}

\item{phi1}{decay parameter. Control the strength of the covariance between the points.}

\item{phi2}{decay parameter. Control the strength of the covariance between the points.}

\item{theta}{angle of the anisotropic field.}

\item{coords}{matrix holding the coordinates of the points. Remember that the points must be projected if you want a correaltion matrix that make sense.}
}
\value{
A Gaussian anisotropic covariance matrix between points.
}
\description{
Gaussian anisotropic covariance function
}
