% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_iso_gaus.R
\name{covar_iso_gaus}
\alias{covar_iso_gaus}
\title{Gaussian isotropic covariance function}
\usage{
covar_iso_gaus(sigma = 1.25, eta = 0.01, phi = 5, coords = expand.grid(x
  = seq(-10, 10, by = 2), y = seq(-10, 10, by = 2)))
}
\arguments{
\item{sigma}{standard deviation for the covariance function}

\item{eta}{smoothness parameters. Control how smooth the correlation function is.}

\item{phi}{decay parameters. Control the strength of the correlation between the points.}

\item{coords}{matrix holding the coordinates of the points.}
}
\value{
A Gaussian isotropic covariance matrix between points.
}
\description{
Calulate a Gaussian covariance functions based on a vector of coordinates.Remember that the points must be projected if you want a correaltion matrix that make sense.
}
