% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dirs.R
\name{create_dirs}
\alias{create_dirs}
\title{Create directories}
\usage{
create_dirs(other_dirs = NULL)
}
\arguments{
\item{other_dirs}{If needed, a character vector containing the names for extra directories.}
}
\description{
Creates a set of standard directories: raw_data, data, rmd, docs, scripts
}
\examples{

\dontrun{
create_dirs() # This creates default directories

create_dirs(other_dirs = "figures") # Creates all default directories and a figures directory
}


}
