% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_emlab_dirs.R
\name{create_emlab_dirs}
\alias{create_emlab_dirs}
\title{Create emLab directories}
\usage{
create_emlab_dirs(project_codename, other_dirs = NULL)
}
\arguments{
\item{project_codename}{The code name used for the project. It should match the name used in the "current project" directory at emLab}

\item{other_dirs}{If needed, a character vector containing the names for extra directories.}
}
\description{
Creates a set of standard directories in the emLab "current projects" foler: raw_data, processed_data, output_data
}
\examples{

\dontrun{
create_dirs() # This creates default directories

create_dirs(other_dirs = "reports") # Creates all default directories and an "reports" directory
}


}
