% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_manuscript.R
\name{create_manuscript}
\alias{create_manuscript}
\title{Create manuscript}
\usage{
create_manuscript(type = "html")
}
\arguments{
\item{type}{A string that specifies the type of manuscript between \code{html_document} (the default) and \code{pdf_document}}
}
\description{
Create manuscript
}
