% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadataB.R
\name{create_metadataB}
\alias{create_metadataB}
\title{Create a bare metadata file from a dataframe that you plan to save as a csv file}
\usage{
create_metadataB(x, file = NULL, path = ".")
}
\arguments{
\item{x}{dataframe in your R environment that you'd like to create a metadata file for}

\item{file}{string for the filename that you plan to save (must match the filename you used to save the .csv but sans the ".csv" bit)}

\item{path}{directory for where the readme will be saved) (e.g. "Downloads/")}
}
\description{
Generates a metadata file that matches the emLab SOP that is completely bare apart from the absolute basics
}
