% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_readme.R
\name{create_readme}
\alias{create_readme}
\title{Create README}
\usage{
create_readme(repo = "Repository")
}
\arguments{
\item{repo}{(character) The repository name, or any other title, defaults to 'Repository'}
}
\description{
Creates a README.md file containing the structure of the repository and a footer
}
\examples{

\dontrun{
create_readme() # Creates a readme file
}

}
