% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_setup_script.R
\name{create_setup_script}
\alias{create_setup_script}
\title{Create setup script}
\usage{
create_setup_script(project_codename)
}
\arguments{
\item{project_codename}{The code name used for the project. It should match the name used in the "current project" directory at emLab}
}
\description{
Create setup script
}
