% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Coefficient of variation}
\usage{
cv(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or an R object which is coercible to one by as.double(x)}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
the coeficient of variation
}
\description{
Coefficient of variation
}
\examples{
foo <- rnorm(100,10,10)
sd(foo)/mean(foo)
cv(foo)
}
