% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_to_rad.R
\name{deg_to_rad}
\alias{deg_to_rad}
\title{Title Degree angle to radian angle}
\usage{
deg_to_rad(x)
}
\arguments{
\item{x}{angle value in degrees}
}
\value{
angle value in radian
}
\description{
Transform degree angle to radian angle
}
\examples{
 deg_to_rad(90)
#END
}
