% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_test.R
\name{f_test}
\alias{f_test}
\title{f test for Bayesian parameters estimates}
\usage{
f_test(x)
}
\arguments{
\item{x}{vector holding the parameter draw from a MCMC of HMCMC sampler}
}
\value{
the proportion of the parmater estimate that is on the same side of the mean.
}
\description{
Check what proportion of a parameter estimate fall on the same side of the mean.  This value can be use as reference of how confidnet your estimate of the parameter is.
}
\examples{
f_test(rnorm(1e5,mean=-3,sd=6)) #low confidence
f_test(rnorm(1e5,mean=-3,sd=2)) #high confidence
f_test(rnorm(1e5,mean=-3,sd=0.5)) #Certain
#END
}

