% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_commits.R
\name{get_commits}
\alias{get_commits}
\title{Get git commits}
\usage{
get_commits(
  path = "~/Documents/GitHub/",
  target_user,
  from = "2019-01-01",
  export = F,
  filter = F
)
}
\arguments{
\item{path}{Path to the folder where you keep your repos (usually Documents/GitHub/)}

\item{target_user}{Github username(s)}

\item{from}{Initial date}

\item{export}{Should the function return a value (de default) or export the data (set this to TRUE)}
}
\value{
Either a data.frame object or nothing
}
\description{
This function goes into your github repos and extracts the git commits
}
\examples{

get_commits(path = "", target_user = "jcvdav", from = "2019-01-01", export = F)

}
