% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filestream_path.R
\name{get_filestream_path}
\alias{get_filestream_path}
\title{Get Filestream path}
\usage{
get_filestream_path()
}
\value{
The OS-specific mounting of FIleStream
}
\description{
Identifies the operative system in which we are running, and returns the correct path to where Filestream is mounted
}
