% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Get tables from GoogleBigQuery}
\usage{
get_table(project = "ucsb-gfw", dataset = NULL, table = NULL,
  allowLargeResults = TRUE, show_tables = TRUE)
}
\arguments{
\item{project}{Name of the project, defaults to \code{ucsb-gfw}}

\item{dataset}{Name of the dataset where the table lives}

\item{table}{Name of the table to download}

\item{allowLargeResults}{Allow large results?}

\item{show_tables}{Show all tables within the dataset before downloading?}
}
\value{
An object of clas \code{tbl.df}
}
\description{
This function downloads tables from GoogleBigQuery and brings them into memory as a normal r data.frame. The first time it is ran, you will have to authenticate into your GBQ account.
}
\examples{

\dontrun{
my_data <- get_table(dataset = "foreign_fishing_ren",
                     table = "annual_foreign_fishing")
}

}
