% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme_plot.R
\name{ggtheme_plot}
\alias{ggtheme_plot}
\title{Plot theme}
\usage{
ggtheme_plot(font_size = 10, font_family = "", line_size = 0.5)
}
\arguments{
\item{font_size}{The size of the font for \code{element_text()}. Defaults to \code{font_size = 10}.}
}
\description{
Creates a standard theme for all my plots
}
