% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilogit.R
\name{ilogit}
\alias{ilogit}
\title{Logit and inverse logit}
\usage{
ilogit(x)
}
\arguments{
\item{x}{a real number}
}
\value{
a nuber between 0 and 1
}
\description{
logit compute the logit function while ilogit estimate the  inverse-logit of a number
}
\examples{
logit(0.90)
ilogit(2.20)
#END
}

