% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_lazy.R
\name{im_lazy}
\alias{im_lazy}
\title{Lazy repository settup}
\usage{
im_lazy(first_doc = FALSE, ...)
}
\description{
This is a wrapper arround \code{create_dirs} and \code{create_readme} that performs both opperations
}
