% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_ggsave.R
\name{lazy_ggsave}
\alias{lazy_ggsave}
\title{Save a ggplot object as a figure, with some defaults (PNG and PDF only)}
\usage{
lazy_ggsave(plot, filename, width = 14, height = 10)
}
\arguments{
\item{plot}{A ggplot object}

\item{filename}{Filename (with no extensions!)}

\item{width}{Width of the figure (in cm)}

\item{height}{Height of the figure (in cm)}
}
\value{
Does not return anything, just saves a file to disk
}
\description{
Save a ggplot object as a figure, with some defaults (PNG and PDF only)
}
