% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Logit and inverse logit}
\usage{
logit(x)
}
\arguments{
\item{x}{a nuber between 0 and 1}
}
\value{
a real number
}
\description{
logit compute the logit function while ilogit estimate the  inverse-logit of a number
}
\examples{
logit(0.90)
ilogit(2.20)
#END
}
