% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_regtable.R
\name{my_regtable}
\alias{my_regtable}
\title{Customized \code{stargazer} regression table}
\usage{
my_regtable(
  model_list,
  type = "latex",
  single.row = T,
  omit.stat = c("adj.rsq", "ser"),
  ...
)
}
\arguments{
\item{model_list}{A list of models}

\item{type}{A character vector specifying if the table should be latex or html}

\item{single.row}{A logical indicating if standard errors should be shown next to coefficients instead of below}

\item{omit.stat}{Regression statistics to omit from the table. For a full list, see stargazer::stargazer_stat_code_list}

\item{...}{Other parameters to pass to \code{stargazer}}
}
\description{
Creates a stargazer table that fits my prefered design
}
\examples{

\dontrun{
m1 <- lm(mpg ~ cyl, mtcars)
m2 <- lm(mpg ~ cyl + hp, mtcars)

moels <- list(m1, m2)

my_regtable(model_list = models)

}

}
