% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\alias{normalize_shipname}
\alias{normalize_callsign}
\alias{normalize_economic_unit}
\title{Normalize shipname and callsign}
\usage{
normalize_shipname(name, ...)

normalize_callsign(callsign, ...)

normalize_economic_unit(economic_unit, ...)
}
\arguments{
\item{name}{A character string (or vector of strings) to be normalized}

\item{...}{Any other arguments}

\item{callsign}{A character string (or vector of strings) to be normalized}

\item{economic_unit}{A character string (or vector of strings) to be normalized}
}
\value{
A normmalized shipname or callsign
}
\description{
Performs a series of string normalizations to get a clean shipname or callsign.
This was inspired by Jaeyoon Park's Python Module: \url{https://github.com/jaeyoonpark/shipdataprocess}.

\code{normalize_shipname} will do the following:
\itemize{
  \item Make all letters into capital letters
  \item Remove leading and trailing whitespaces
  \item Remove linebreaks
  \item Remove all references to vessel codes (\emph{i.e.} \code{MFV} stands for Marine Fishing Vessel)
  \item Convert numbers expressed as letters to numbers (\emph{i.e.} \code{ONE} becomes \code{1})
  \item Convert roman numerals to arabic numbers
  \item Remove \code{NO.} when it references a number (\emph{i.e.} \code{BOAT NO.5} becomes \code{BOAT5})
  \item Remove all special characters
}

\code{normalize_callsign} will:
\itemize{
  \item Make all letters into capital letters
  \item Remove leading and trailing whitespaces
  \item Remove linebreaks
  \item Remove leading zeroes (\emph{i.e.} \code{007JAMESBOND} becomes \code{7JAMESBOND})
  \item Remove all special characters
}
}
\examples{
# Normalize a shipname that contains special
# characters and roman numerals.
library(startR)
shipname <- "weird-+\%()<>$;!&'`#/boat name IV"
normalize_shipname(shipname)

# Normalize a callsign starting with zeros
# and weird characters
callsign <- "0020300a-+\%()<>$;!&'`#/"
normalize_callsign(callsign)
}
