% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_shipname.R
\name{normalize_shipname}
\alias{normalize_shipname}
\title{Normalize shipname}
\usage{
normalize_shipname(name)
}
\arguments{
\item{name}{A character string (or vecctor of strings) to be normalized}
}
\value{
A normmalized name
}
\description{
Performs a series of string normalizations to get a clean shipname
}
