% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_iris.R, R/normalize.R, R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\alias{new_iris}
\alias{normalize_shipname}
\alias{normalize_callsign}
\title{New iris}
\usage{
new_iris(...)

lhs \%>\% rhs
}
\arguments{
\item{...}{Ignores all parameters}

\item{name}{A character string (or vector of strings) to be normalized}

\item{callsign}{A character string (or vector of strings) to be normalized}

\item{...}{Any other arguments}
}
\value{
A normmalized shipname or callsign
}
\description{
Creates a set of 50 lines with data from Ryan Anderson: https://dreamtolearn.com/ryan/data_analytics_viz/18, who made up some 50 new values for a fictional species: Iris mythica

Performs a series of string normalizations to get a clean shipname or callsign.
This was inspired by Jaeyoon Park's Python Module \url{https://github.com/jaeyoonpark/shipdataprocess}.

See \code{magrittr::\link[magrittr]{\%>\%}} for details.
}
\examples{
# Normalize a shipname that contains special
# characters and roman numerals.
shipname <- "weird -+\%()<>$;!&'`\\\\.#/boat name IV"
normalize_shipname(shipname)

# Normalize a callsign starting with zeros
# and weird characters
callsign <- "0020300a-+\%()<>$;!&'`\\\\.#/"
normalize_callsign(callsign)
}
\keyword{internal}
