% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad_to_deg.R
\name{rad_to_deg}
\alias{rad_to_deg}
\title{Title Radial angle to degree angle}
\usage{
rad_to_deg(x)
}
\arguments{
\item{x}{angle value in radian}
}
\value{
angle value in degrees
}
\description{
Transform radian angle to degree angle
}
\examples{
 rad_to_deg(2*pi)
#END
}
