% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_doc.R
\name{render_doc}
\alias{render_doc}
\title{Render Document}
\usage{
render_doc(file, rmd_dir = "rmd", output_dir = "docs", ...)
}
\arguments{
\item{file}{A character vector specifying the file name}

\item{rmd_dir}{The directory where the file lives, specified as a character vector}

\item{output_dir}{The output directory, where the knitted versions will be placed}
}
\description{
Calls \code{rmarkdown::render()} to produce the output of a given rmd file
}
\examples{

I have a file called \\code{"manuscript.rmd"} that lives in my \\code{"repo/rmd/"} directory.
\dontrun{
render_doc(file = "manuscript.rmd")
}

}
