% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_rotate.R
\name{st_rotate}
\alias{st_rotate}
\title{Rotate an sf object}
\usage{
st_rotate(x)
}
\arguments{
\item{x}{an sf object}
}
\value{
the same object, with coordinates rotated
}
\description{
This takes an sf object with coordinates in -180 to 180
and converts it to an object 0 to 360 (like the ones used in climate
model outpt)
}
