% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_err.R
\name{std_err}
\alias{std_err}
\title{Standard error}
\usage{
std_err(x)
}
\arguments{
\item{a}{numeric vector or an R object which is coercible to one by as.double(x)}
}
\value{
a real number representing the standard error
}
\description{
Standard error
}
\examples{
x <- rnorm(100)
mean(x)
std_err(x)
#End
}

