% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transliterations_all.R
\docType{data}
\name{transliterations_all}
\alias{transliterations_all}
\title{transliterations_all}
\format{
A data frame with 38883 rows and 3 variables:
}
\source{
\url{https://github.com/rich-iannone/UnidecodeR/blob/master/data/transliterations_all.rda}
}
\usage{
transliterations_all
}
\description{
A dataset containing transliterations
}
\keyword{datasets}
