% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usual_suspects.R
\name{usual_suspects}
\alias{usual_suspects}
\title{Package loading}
\usage{
usual_suspects(pkgs = NULL)
}
\arguments{
\item{pkgs}{A character vector containing the packages that should be loaded. The fedault ones include \code{janitor}, \code{here}, \code{sf}, and \code{tidyverse}}
}
\description{
Loads the usual packages from my every-day workflow
}
