% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_all.R
\name{create_all}
\alias{create_all}
\title{Create all}
\usage{
create_all(local_dirs = T, emlab_dirs = T, setup_script = T, readme = T)
}
\arguments{
\item{readme}{}
}
\description{
This is a wrapper around \code{create_local_dirs}, \code{create_emlab_dirs}, \code{create_setup_script} and \code{create_readme}
}
