% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_readme.R
\name{create_readme}
\alias{create_readme}
\title{Creat README}
\usage{
create_readme(repo = "Repository",
  footer = "<a href=\\"https://orcid.org/0000-0003-1245-589X\\" target=\\"orcid.widget\\" rel=\\"noopener noreferrer\\" style=\\"vertical-align:top;\\"><img src=\\"https://orcid.org/sites/default/files/images/orcid_16x16.png\\" style=\\"width:1em;margin-right:.5em;\\" alt=\\"ORCID iD icon\\">orcid.org/0000-0003-1245-589X</a>")
}
\arguments{
\item{repo}{(character) The repository name, or any other title, defaults to 'Repository'}

\item{footer}{(character) Text to be included in the footer. Defaults to my orcid}
}
\description{
Creates a README.md file containing the structure of the repository and a footer
}
\examples{

\dontrun{
create_readme() # Creates a readme file
}

}
