% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_iris.R, R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\alias{new_iris}
\alias{\%>\%}
\title{New iris}
\usage{
new_iris(...)

lhs \%>\% rhs
}
\arguments{
\item{...}{Ignores all parameters}
}
\description{
Creates a set of 50 lines with data from Ryan Anderson: https://dreamtolearn.com/ryan/data_analytics_viz/18, who made up some 50 new values for a fictional species: Iris mythica

See \code{magrittr::\link[magrittr]{\%>\%}} for details.
}
\keyword{internal}
