% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_as_cols.R
\name{sfc_as_cols}
\alias{sfc_as_cols}
\title{sfc_as_cols}
\usage{
sfc_as_cols(x, names = c("lon", "lat"))
}
\arguments{
\item{x}{An object of class df sf}

\item{names}{}
}
\value{
a data.frame
}
\description{
Take the geometry column of an sf point object and turn it into columns of the data.frame
}
